back to the default leader
I tried the spacebar for a while, but I keep mis-hitting it. My fingers are used to putting a space *after* the things I write, not in front of it. Backtick works better for me, and also: the placement of the backtick on a Dutch keyboard is way better, so there isn’t a good reason not to use it.

I am curious about this space-as-command thing I came up with. Worth trying out.