fix spawning vim folder
The problem here is that ~/.vim is a folder. That way, if you run the script for the second time, it adds the symlink to that folder, which is undesired. The nice way about this `rm` is: it will error if ~/.vim is a regular folder (instead of a symlink).