a more terminal based git workflow
As I use Vim as my editor, I work mainly from the terminal. For some
reason however, I never really got comfortable with Git on the terminal.
I still use a GUI for it: first Tower and now Fork. This all happened
because I did not know how to stage separate lines of a file, which I
could do in the GUI but not on the command line (there is a solution
though: `git add -p`, among others).

I stil dread merging and rebasing and merge conflicts in the terminal,
but with Fugitive and my knowledge of Git internals I should be fine
nowadays.