add new Tmux switcher
This one just lists all folders in ~/code/ and looks for a file named
.tmux-session-name in the root. The contents of this file are used to
name the session.

If the session does not exist, it is created with that filename, with
the folder opened. If it does exist we just switch to it.

Added the .tmux-session-name to the .gitignore_global, so it will never
be added to these git repos.

Also adds a special case for the dotfiles, as I don't store them in my
~/code/ folder but directly as ~/dotfiles.