fix ctrl+J and ctrl+K in fzf screens
Specifically, I use fzf inside of Vim as my main way to get around in a
project (by searching for filenames). But to get up and down in that
list, fzf uses Ctrl + J and K, which were also captured by vim-tmux-
navigator. I never actually use the terminal buffers in Vim, so I can
get rid of `tnoremap`, and tweaking the regex makes sure Tmux sends the
keys to fzf as well (as it would with Vim).

This is a fix taken from the uncommitted changes in my mac's dotfiles.